[ncl] fix web errors
