+++ apps/native-component-list/src/navigation/RootNavigation.tsx
@@ -3,6 +3,7 @@ import { createStackNavigator } from '@react-navigation/stack';
 import * as Linking from 'expo-linking';
 import * as React from 'react';
 import { Text } from 'react-native';
+import { GestureHandlerRootView } from 'react-native-gesture-handler';
 
 import RedirectScreen from '../screens/RedirectScreen';
 import SearchScreen from '../screens/SearchScreen';
@@ -33,16 +34,18 @@ export const linking: LinkingOptions<object> = {
 
 export default function RootNavigation() {
   return (
-    <NavigationContainer linking={linking} fallback={<Text>Loading…</Text>}>
-      <Switch.Navigator screenOptions={{ presentation: 'modal', headerShown: false }}>
-        <Switch.Screen name="main" component={MainTabNavigator} />
-        <Switch.Screen name="redirect" component={RedirectScreen} />
-        <Switch.Screen
-          name="searchNavigator"
-          component={SearchScreen}
-          options={{ cardStyle: { backgroundColor: 'transparent' } }}
-        />
-      </Switch.Navigator>
-    </NavigationContainer>
+    <GestureHandlerRootView style={{ flex: 1 }}>
+      <NavigationContainer linking={linking} fallback={<Text>Loading…</Text>}>
+        <Switch.Navigator screenOptions={{ presentation: 'modal', headerShown: false }}>
+          <Switch.Screen name="main" component={MainTabNavigator} />
+          <Switch.Screen name="redirect" component={RedirectScreen} />
+          <Switch.Screen
+            name="searchNavigator"
+            component={SearchScreen}
+            options={{ cardStyle: { backgroundColor: 'transparent' } }}
+          />
+        </Switch.Navigator>
+      </NavigationContainer>
+    </GestureHandlerRootView>
   );
 }
